1. Fixed a bug in determining the number of DFFs required when the source and target of a path are respectively an AS gate and an SA gate; 2. Implemented a new T1 detection algorithm, which creates nodes in the KLUT for the detected T1 cells; 3. Added support for the new implementation, such as invoking ‘is_dangling’ check when calling ‘foreach_node’ and ‘foreach_fanin’; 4. Replaced the explicit usages of ‘uint32_t’ with ‘uint64_t’ for ‘klut::node’ and ‘klut::signal’ related cases.
