+++ experiments/skip/mph_cpsat_FA.cpp
@@ -36,9 +36,6 @@
 #include <lorina/bench.hpp>
 #include <mockturtle/algorithms/rsfq/rsfq_network_conversion.hpp>
 #include <mockturtle/algorithms/rsfq/rsfq_path_balancing.hpp>
-
-#include <mockturtle/algorithms/multiphase.hpp>
-
 #include <mockturtle/algorithms/mapper.hpp>
 #include <mockturtle/algorithms/nodes.hpp>
 #include <mockturtle/algorithms/node_resynthesis.hpp>
@@ -79,8 +76,6 @@ typedef mockturtle::xmg_network   xmg;
 typedef mockturtle::mig_network   mig;
 typedef mockturtle::aig_network   aig;
 typedef uint64_t node_t;
-template <size_t N, typename T>
-using array_map = phmap::flat_hash_map<std::array<uint32_t, N>, T, ArrayHash<N>>;
 
 // Sunmagnetics Technology Library
 constexpr std::array<int, 13> COSTS_MAP = {7, 9, 8, 8, 12, 8, 999, 999, 999, 8, 3, 0, 25};
@@ -183,6 +178,25 @@ std::tuple<mockturtle::binding_view<klut>, mockturtle::map_stats, double, double
   return std::make_tuple( res, st, total_ndff, total_area, cec );
 }
 
+constexpr uint8_t PI_GATE = 0u;
+constexpr uint8_t AA_GATE = 1u;
+constexpr uint8_t AS_GATE = 2u;
+constexpr uint8_t SA_GATE = 3u;
+constexpr uint8_t FA_GATE = 4u;
+
+union NodeData 
+{
+  struct 
+  {
+    unsigned int sigma : 29;
+    unsigned int type : 3;
+  };
+  uint32_t value;
+  NodeData() : value(0) {}
+  NodeData(uint32_t _sigma, uint8_t _type) : sigma(_sigma), type(_type) {}
+  NodeData(uint32_t _value) : value(_value) {}
+};
+
 const std::vector<std::string> GATE_TYPE { "PI", "AA", "AS", "SA", "FA" }; //, "PO"
 
 typedef uint32_t glob_phase_t;
@@ -224,6 +238,252 @@ public:
       fmt::join(primitive.fanins, ", "), primitive.is_spl);
   }
 };
+
+// #if false
+std::tuple<klut, std::unordered_map<klut::signal, Primitive<klut>>, int64_t> decompose_to_klut(mockturtle::binding_view<klut> src, phmap::flat_hash_map<ULL, Node> nodemap, phmap::flat_hash_map<std::string, LibEntry> entries, bool verbose = false)
+{
+  /* record the information of each signal in tgt, including: signal index (why?), */
+  /* truth table ( represented using an uint16 ), type ( PI/AA/AS/AS ), fan-ins,   */
+  /* and if it is a splitter                                                       */
+  std::unordered_map<klut::signal, Primitive<klut>> tgt_sig_params;
+  /* record the correspondence between nodes in src and node in tgt                */
+  std::unordered_map<klut::signal, klut::signal> src2tgt;
+  int64_t area = 0;
+
+  /* src: a 4-LUT with gates in the target technology libarary binded to each node */
+  /* tgt: a 2-LUT ( essentially 4-LUT ), each node is associated with a NodeData,  */
+  /* which contains the assigned stage and the gate type ( PI/SA/AS/AA ), denoted  */
+  /* using 'value' since src is converted into tgt by decomposing each 4-LUT into  */
+  /* 2-LUTs, for each node in src, there is a node in tgt implementing the same    */
+  /* function                                                                      */
+  klut tgt;
+  src.foreach_pi( [&]( const klut::signal & src_pi ) 
+  {
+    klut::signal tgt_pi = tgt.create_pi();
+    tgt.set_value(tgt_pi, NodeData(0, PI_GATE).value);
+    src2tgt.emplace(src_pi, tgt_pi);
+    tgt_sig_params.emplace( tgt_pi, Primitive<klut>(tgt_pi, 0xAAAA, AS_GATE, {}));
+  } );
+
+  // The bindings are replaced in forward topological order.
+  std::vector<klut::signal> src_po;
+  src_po.reserve(src.num_pos());
+  src.foreach_po( [&] (auto src_n)
+  {
+    src_po.push_back(src_n);
+  } );
+
+  std::unordered_set<klut::signal> OR_replacement_candidates;
+  std::vector<klut::signal> XOR_gates;
+
+  auto num_node_src = src.size();
+  auto ctr = 0u;
+  mockturtle::topo_view<klut>( src ).foreach_node( [&]( auto src_node ) 
+  {
+    DEBUG_PRINT("Processing node {0} ({1} out of {2})\r", src_node, ++ctr, num_node_src);
+    if (ctr == num_node_src)
+    {
+      DEBUG_PRINT("\n");
+    }
+    if ( !src.has_binding( src_node ) )
+    {
+      return;
+    }
+
+    /* get gate type in the technology library */
+    auto const& g = src.get_binding( src_node );
+
+    // handing constants
+    if (g.name == "one" && g.expression == "CONST1")
+    {
+      klut::signal tgt_sig = tgt.get_constant( true );
+      src2tgt.emplace(src_node, tgt_sig);
+      return;
+    }
+    else if (g.name == "zero" && g.expression == "CONST0")
+    {
+      klut::signal tgt_sig = tgt.get_constant( false );
+      src2tgt.emplace(src_node, tgt_sig);
+      return;
+    }
+    
+    /* for each cell in the technology library, there is a primitive gates-based     */
+    /* implementation, prepared in advance and ready for a look-up                   */
+    LibEntry entry = entries.at(g.name);
+    Node & root = nodemap[entry.hash];
+
+    // fmt::print("Binding : {0}\n", entry.to_str());
+    // fmt::print("NODE function : 0x{0:{2}x}, {1}\n", root.func, root.to_str(), std::max(1 << (src.fanin_size(src_node) - 2), 1) );
+
+    // kitty::dynamic_truth_table func(src.fanin_size(src_node));
+    // func._bits = root.func;
+    // tgt._create_node(parent_tgt_1, parent_tgt_2);
+
+    // _create_node( { a }, 2 )
+
+    // Get the topological order of internal nodes (i.e., nodes inside the cell)
+    std::vector<ULL> topo_order;
+    root.topo_sort(nodemap, topo_order, true);
+
+    std::vector<klut::signal> tgt_fanins;
+    src.foreach_fanin(src_node, [&](const auto & src_fanin)
+    {
+      tgt_fanins.push_back(src2tgt.at(src_fanin));
+      // fmt::print("\tRecorded fanin of src_node {0}:\n\t\tsrc_fanin: {1}\n\t\ttgt_fanin: {2}\n", src_node, src_fanin, tgt_fanins.back().data);
+    } );
+
+    std::unordered_map<ULL, klut::signal> node2tgt;
+
+    /* for each cell/compound gate in src, implement it using primitive gates in tgt */
+    for (ULL hash : topo_order)
+    {
+      Node & node = nodemap.at(hash);
+      if (node.last_func == fPI)
+      {
+        // determine which index of the PI is needed 
+        char letter = node.pi_letter();
+
+        auto pi_idx = std::find(entry.chars.begin(), entry.chars.end(), letter) - entry.chars.begin();
+        node2tgt.emplace(hash, tgt_fanins[pi_idx]);
+        
+        // tgt_sig_params.emplace();
+      }
+      else if (node.last_func == fDFF)
+      {
+        klut::signal parent_tgt = node2tgt.at(node.parent_hashes.front());
+        node2tgt.emplace(hash, parent_tgt);
+        // tgt_sig_params.emplace(tgt_sig, Primitive<klut>(tgt_sig, 0xAAAAu, 1u, { parent_tgt }));
+        //  nothing else to do, no new constraints, no contribution to objective function
+      }
+      else if (node.last_func == fNOT)
+      {
+        klut::signal parent_tgt = node2tgt.at(node.parent_hashes.front());
+        klut::signal tgt_sig = tgt.create_not( parent_tgt );
+        tgt.set_value(tgt_sig, NodeData(0, AS_GATE).value);
+        node2tgt.emplace(hash, tgt_sig);
+        tgt_sig_params.emplace(tgt_sig, Primitive<klut>(tgt_sig, 0x5555u, AS_GATE, { parent_tgt }));
+        area += COSTS_MAP[fNOT];
+        DEBUG_PRINT("ADDED NOT = {}\n", area);
+      }
+      else if (node.last_func == fAND)
+      {
+        klut::signal parent_tgt_1 = node2tgt.at(node.parent_hashes.front());
+        klut::signal parent_tgt_2 = node2tgt.at(node.parent_hashes.back());
+        klut::signal tgt_sig = tgt.create_and(parent_tgt_1, parent_tgt_2);
+        tgt.set_value(tgt_sig, NodeData(0, SA_GATE).value);
+        // fmt::print("Created node n{0} = AND({1}, {2})\n", tgt_sig, parent_tgt_1, parent_tgt_2);
+        node2tgt.emplace(hash, tgt_sig);
+        tgt_sig_params.emplace(tgt_sig, Primitive<klut>(tgt_sig, 0x8888u, SA_GATE, { parent_tgt_1, parent_tgt_2 }));
+        area += COSTS_MAP[fAND];
+        DEBUG_PRINT("ADDED AND = {}\n", area);
+      }
+      else if (node.last_func == fOR) 
+      {
+        klut::signal parent_tgt_1 = node2tgt.at(node.parent_hashes.front());
+        klut::signal parent_tgt_2 = node2tgt.at(node.parent_hashes.back());
+        klut::signal tgt_sig = tgt.create_or(parent_tgt_1, parent_tgt_2);
+        tgt.set_value(tgt_sig, NodeData(0, SA_GATE).value);
+        // fmt::print("Created node n{0} = OR({1}, {2})\n", tgt_sig, parent_tgt_1, parent_tgt_2);
+        node2tgt.emplace(hash, tgt_sig);
+        tgt_sig_params.emplace(tgt_sig, Primitive<klut>(tgt_sig, 0xEEEEu, SA_GATE, { parent_tgt_1, parent_tgt_2 }));
+        OR_replacement_candidates.emplace(tgt_sig);
+        area += COSTS_MAP[fOR];
+        DEBUG_PRINT("ADDED OR  = {}\n", area);
+      }
+      else if (node.last_func == fCB)
+      {
+        klut::signal parent_tgt_1 = node2tgt.at(node.parent_hashes.front());
+        klut::signal parent_tgt_2 = node2tgt.at(node.parent_hashes.back());
+        klut::signal tgt_sig = tgt.create_or(parent_tgt_1, parent_tgt_2);
+        tgt.set_value(tgt_sig, NodeData(0, AA_GATE).value);
+        // fmt::print("Created node n{0} = CB({1}, {2})\n", tgt_sig, parent_tgt_1, parent_tgt_2);
+        node2tgt.emplace(hash, tgt_sig);
+        tgt_sig_params.emplace(tgt_sig, Primitive<klut>(tgt_sig, 0xEEEEu, AA_GATE, { parent_tgt_1, parent_tgt_2 }));
+        area += COSTS_MAP[fCB];
+        DEBUG_PRINT("ADDED CB  = {}\n", area);
+      }
+      else if (node.last_func == fXOR)
+      {
+        klut::signal parent_tgt_1 = node2tgt.at(node.parent_hashes.front());
+        klut::signal parent_tgt_2 = node2tgt.at(node.parent_hashes.back());
+        klut::signal tgt_sig = tgt.create_xor(parent_tgt_1, parent_tgt_2);
+        tgt.set_value(tgt_sig, NodeData(0, AS_GATE).value);
+        // fmt::print("Created node n{0} = XOR({1}, {2})\n", tgt_sig, parent_tgt_1, parent_tgt_2);
+        node2tgt.emplace(hash, tgt_sig);
+        tgt_sig_params.emplace(tgt_sig, Primitive<klut>(tgt_sig, 0x6666u, AS_GATE, { parent_tgt_1, parent_tgt_2 }));
+        XOR_gates.push_back(tgt_sig);
+        area += COSTS_MAP[fXOR];
+        DEBUG_PRINT("ADDED XOR = {}\n", area);
+      }
+      // else if 
+      else
+      {
+        throw "Unsupported function";
+      }
+    }
+    ULL root_hash = topo_order.back();
+    klut::signal tgt_sig = node2tgt.at(root_hash);
+    src2tgt.emplace(src_node, tgt_sig);
+  } );
+
+  for (klut::signal const & xor_sig : XOR_gates)
+  {
+    //std::vector<klut::signal> stack;
+    //std::vector<klut::signal> visited;
+    Primitive<klut> & xor_prim = tgt_sig_params.at(xor_sig);
+
+    for (klut::signal const & fanin : xor_prim.fanins)
+    {
+      // Primitive<klut> & fanin_prim = tgt_sig_params.at(fanin);
+      /* hazzards potentially produced by a CB can lead to incorrect behavior of its   */
+      /* following XOR, such a CB should therefore be replaced by an OR */
+      auto it = std::find(OR_replacement_candidates.begin(), OR_replacement_candidates.end(), fanin);
+      if (it != OR_replacement_candidates.end())
+      {
+        OR_replacement_candidates.erase(it);
+      }
+    }
+  }
+
+  for (klut::signal const & sig : OR_replacement_candidates)
+  {
+    Primitive<klut> & prim = tgt_sig_params.at(sig);
+    tgt.set_value(sig, NodeData(0, AA_GATE).value);
+    prim.type = AA_GATE;
+    area += COSTS_MAP[fCB];
+    area -= COSTS_MAP[fOR];
+    DEBUG_PRINT("REPLACED OR WITH CB = {}\n", area);
+  }
+
+  src.foreach_po([&](auto const & src_po)
+  {
+    tgt.create_po(src2tgt.at(src_po));
+  } );
+  
+  return std::make_tuple(tgt, tgt_sig_params, area);
+}
+
+#if false
+  struct NtkNode
+  {
+    node_t              id;
+    std::vector<node_t> fanins;
+    std::vector<node_t> fanouts;
+    glob_phase_t        phase;
+    uint8_t             type; 
+    klut::signal        ntk_sig;
+    bool                is_spl;
+    bool                is_pi;
+    bool                is_po;
+    bool                is_const;
+    bool                is_valid;
+
+    NtkNode(const klut::signal & _id, const std::vector<klut::signal>& _fanins, const std::vector<klut::signal>& _fanouts, glob_phase_t _phase, uint8_t _type, klut::signal _ntk_sig, bool _is_spl, bool _is_pi, bool _is_const)
+      : id(_id), fanins(_fanins), fanouts(_fanouts), phase(_phase), type(_type), ntk_sig(_ntk_sig), is_spl(_is_spl), is_pi(_is_pi), is_const(_is_const), is_valid(true) {}
+    NtkNode(const klut::signal & _id)
+      : id(_id), fanins({}), fanouts({}), phase(0), type(0), ntk_sig(0), is_spl(false), is_pi(false), is_const(false), is_valid(false) {}
+  };
+#endif
 
 struct Path
 {
@@ -295,43 +555,6 @@ struct Standard_T1_CELL
 
   Standard_T1_CELL( const unsigned in_phase_, const uint8_t sum_truth_table_, const uint8_t carry_truth_table_, const uint8_t cbar_truth_table_ )
     : in_phase( in_phase_ ), sum_truth_table( sum_truth_table_ ), carry_truth_table( carry_truth_table_ ), cbar_truth_table( cbar_truth_table_ ) {}
-
-  /// @brief Checks whether the truth table matches the carry TT of this T1 cell
-  /// @param tt truth table to match
-  /// @param phase - set to *true* if the matched to negated output, or *false* if matched to non-negated output
-  /// @return *true* if there is a match at all
-  bool check_carry( const kitty::dynamic_truth_table & tt, bool & phase ) const
-  {
-    if ( tt._bits[0] == ~carry_truth_table )
-    {
-      phase = true;
-      return true;
-    }
-    else if ( tt._bits[0] != carry_truth_table )
-    {
-      return false;
-    }
-    phase = false;
-    return true;
-  }
-  /// @brief Checks whether the truth table matches the cbar TT of this T1 cell
-  /// @param tt truth table to match
-  /// @param phase - set to *true* if the matched to negated output, or *false* if matched to non-negated output
-  /// @return *true* if there is a match at all
-  bool check_cbar( const kitty::dynamic_truth_table & tt, bool & phase ) const
-  {
-    if ( tt._bits[0] == ~cbar_truth_table )
-    {
-      phase = true;
-      return true;
-    }
-    else if ( tt._bits[0] != cbar_truth_table )
-    {
-      return false;
-    }
-    phase = false;
-    return true;
-  }
 };
 
 static std::array<Standard_T1_CELL, 8> standard_T1_cells = 
@@ -342,7 +565,7 @@ static std::array<Standard_T1_CELL, 8> standard_T1_cells =
   Standard_T1_CELL( 3, 0x96, 0x71, 0xf7 ), 
   Standard_T1_CELL( 4, 0x69, 0x8e, 0xef ), 
   Standard_T1_CELL( 5, 0x96, 0x4d, 0xdf ), 
-  Standard_T1_CELL( 6, 0x69, 0xb2, 0xbf ), 
+  Standard_T1_CELL( 6, 0x96, 0x2b, 0xbf ), 
   Standard_T1_CELL( 7, 0x69, 0x17, 0x7f )
 };
 
@@ -434,251 +657,6 @@ struct T1_OUTPUTS
   }
 };
 
-template <typename CutType>
-array_map<3, std::tuple<kitty::dynamic_truth_table, uint32_t>> match_cuts( const std::vector<TT3> & template_tts, const klut & ntk, const CutType & cuts)
-{
-  array_map<3, std::tuple<kitty::dynamic_truth_table, uint32_t>> matching_cuts;
-
-  ntk.foreach_node( [&]( const klut::signal & node ) 
-  { 
-    const auto idx = ntk.node_to_index( node );
-    const auto & node_cut_set = cuts.cuts( idx );
-    // fmt::print("[{}] Extracted {} cuts for node {} (idx={})\n", benchmark, node_cut_set.size(), node, idx);
-    // if (node_cut_set.size() == 0) { return; }
-
-    /* check if the function of the cut is NPN-equivalent to any of the template_tts */
-    for (const auto & cut_entry : node_cut_set)
-    {
-      const auto tt = cuts.truth_table( *cut_entry );
-      if (std::find_if(template_tts.begin(), template_tts.end(), [&](const TT3 & tt){return tt._bits.front() == tt._bits;}) != template_tts.end())
-      {
-        std::array<uint32_t,3> leaves;
-        std::copy(cut_entry->begin(), cut_entry->end(), leaves.begin());
-        matching_cuts.emplace( leaves, std::make_tuple(tt, idx) );
-      }
-    }
-  } );
-  return matching_cuts;
-}
-
-array_map<3, T1_OUTPUTS> find_t1_candidates( 
-  const klut & ntk,
-  const array_map<3, std::tuple<kitty::dynamic_truth_table, uint32_t>>& xor3_cuts,
-  const array_map<3, std::tuple<kitty::dynamic_truth_table, uint32_t>>& maj3_cuts, 
-  const array_map<3, std::tuple<kitty::dynamic_truth_table, uint32_t>>& or3_cuts
-) 
-{
-  array_map<3, T1_OUTPUTS> t1_candidates;
-
-  for ( auto const& [target_leaves, xor3_tt_and_node] : xor3_cuts )
-  {
-    if ( !( maj3_cuts.contains( target_leaves ) &&  // if there's a match with MAJ3
-             or3_cuts.contains( target_leaves ) ) ) // and a match with OR3
-    {
-      continue;
-    }
-    const auto & [xor3_tt, xor_index] = xor3_tt_and_node;
-    const auto & [maj3_tt, maj_index] = maj3_cuts.at(target_leaves);
-    const auto & [ or3_tt,  or_index] = or3_cuts.at(target_leaves); 
-    /* find a candidate potentially using all 3 outputs of an T1 cell */
-    /* check whether the combination of truth table is valid          */
-    for ( auto const& t1_cell : standard_T1_cells )
-    {
-      bool maj_phase{ false }, or_phase{ false };
-      if (
-        xor3_tt._bits[0] != t1_cell.sum_truth_table ||
-        !t1_cell.check_carry(maj3_tt, maj_phase) || 
-        !t1_cell.check_cbar(or3_tt, or_phase)
-      )
-      {
-        continue;
-      }
-
-      /* managed to find a T1 cell-based implementation      */
-      /* figure out the nodes connecting to the output ports */
-      klut::node sum_to       = ntk.index_to_node( xor_index );
-
-      klut::node carry_to     = ntk.get_constant( false );
-      klut::node inv_carry_to = ntk.get_constant( false );
-      config_t1_ports_connection( ntk, maj_phase, maj_index, carry_to, inv_carry_to );
-
-      klut::node cbar_to      = ntk.get_constant( false );
-      klut::node inv_cbar_to  = ntk.get_constant( false );
-      config_t1_ports_connection( ntk,  or_phase, or_index, cbar_to,  inv_cbar_to );
-
-      /* instantiate an T1 cell */
-      t1_candidates.emplace( target_leaves, T1_OUTPUTS( 
-        t1_cell.in_phase, 
-        inv_carry_to != ntk.get_constant( false ), 
-        inv_cbar_to  != ntk.get_constant( false ), 
-        sum_to       != ntk.get_constant( false ), 
-        carry_to     != ntk.get_constant( false ), 
-        cbar_to      == ntk.get_constant( false ), 
-        sum_to, carry_to, inv_carry_to, cbar_to, inv_cbar_to ) );
-      break;
-      /* TODO: think about if it is possible that a design can be implemented under */
-      /* different input phases of a T1 cell; if yes, it is better to enumerate all */
-      /* the potential input phases and choose the one requires the minimum cost    */
-    }
-  }
-
-  for ( auto const& [target_leaves, xor3_tt_and_node] : xor3_cuts )
-  {
-    if ( !maj3_cuts.contains( target_leaves ) ||   // if there's no XOR-MAJ match
-          t1_candidates.contains( target_leaves ) ) // or a 3-way match already found
-    {
-      continue;
-    }
-
-    auto [xor3_tt, xor_index] = xor3_tt_and_node;
-    auto [maj3_tt, maj_index] = maj3_cuts.at(target_leaves);
-    /* find a candidate potentially using 2 outputs of an T1 cell  */
-    /* ( XOR3, MAJ3 ), check whether the combnation of tt is valid */
-    bool maj_phase{ false };
-    for ( auto const& t1_cell : standard_T1_cells )
-    {
-      if ( xor3_tt._bits[0] != t1_cell.sum_truth_table )
-      {
-        continue;
-      }
-
-      bool maj_phase{ false };
-      if (!t1_cell.check_carry(maj3_tt, maj_phase))
-      {
-        continue;
-      }
-
-      /* managed to find a T1 cell-based implementation      */
-      /* figure out the nodes connecting to the output ports */
-      klut::node sum_to       = ntk.index_to_node( xor_index );
-      klut::node carry_to     = ntk.get_constant( false );
-      klut::node inv_carry_to = ntk.get_constant( false );
-      config_t1_ports_connection( ntk, maj_phase, maj_index, carry_to, inv_carry_to );
-
-      /* instantiate an T1 cell */
-      t1_candidates.emplace( target_leaves, T1_OUTPUTS( 
-        t1_cell.in_phase, 
-        inv_carry_to != ntk.get_constant( false ), 
-        false, 
-        true, 
-        carry_to != ntk.get_constant( false ), 
-        false, 
-        sum_to, 
-        carry_to, 
-        inv_carry_to, 
-        ntk.get_constant( false ), 
-        ntk.get_constant( false ) ) );
-      break;
-    }
-  }
-
-  for ( auto const& [target_leaves, xor3_tt_and_node] : xor3_cuts )
-  {
-    if ( !or3_cuts.contains( target_leaves ) ||  // if there's no match with OR3
-        t1_candidates.contains( target_leaves )) // or a 3-way match already found
-    {
-      continue;
-    }
-    auto [xor3_tt, xor_index] = xor3_tt_and_node;
-    auto [or3_tt, or_index] = or3_cuts.at(target_leaves); 
-    /* find a candidate potentially using 2 outputs of an T1 cell  */
-    /* ( XOR3,  OR3 ), check whether the combnation of tt is valid */
-    bool or_phase{ false };
-    for ( auto const& t1_cell : standard_T1_cells )
-    {
-      if ( xor3_tt._bits[0] != t1_cell.sum_truth_table )
-      {
-        continue;
-      }
-
-      bool or_phase{ false };
-      if (!t1_cell.check_cbar(or3_tt, or_phase))
-      {
-        continue;
-      }
-
-      /* managed to find a T1 cell-based implementation      */
-      /* figure out the nodes connecting to the output ports */
-      klut::node sum_to      = ntk.index_to_node( xor_index );
-      klut::node cbar_to     = ntk.get_constant( false );
-      klut::node inv_cbar_to = ntk.get_constant( false );
-      config_t1_ports_connection( ntk, or_phase, or_index, cbar_to, inv_cbar_to );
-
-      /* instantiate an T1 cell */
-      t1_candidates.emplace( target_leaves, T1_OUTPUTS( 
-        t1_cell.in_phase, 
-        false, 
-        inv_cbar_to != ntk.get_constant( false ), 
-        true, 
-        false, 
-        cbar_to != ntk.get_constant( false ), 
-        sum_to, 
-        ntk.get_constant( false ), 
-        ntk.get_constant( false ), 
-        cbar_to, 
-        inv_cbar_to ) );
-      break;
-    }
-  }
-
-  for ( auto const& [target_leaves, maj3_tt_and_node] : maj3_cuts )
-  {
-    if ( !or3_cuts.contains( target_leaves ) ||  // if there's no match with OR3
-        t1_candidates.contains( target_leaves )) // or a 3-way match already found
-    {
-      continue;
-    }
-    auto [maj3_tt, maj_index] = maj3_tt_and_node;
-    auto [or3_tt, or_index] = or3_cuts.at(target_leaves); 
-    /* find a candidate potentially using 2 outputs of an T1 cell  */
-    /* ( MAJ3,  OR3 ), check whether the combnation of tt is valid */
-    for ( auto const& t1_cell : standard_T1_cells )
-    {
-      bool maj_phase{ false }, or_phase{ false };
-      if ( maj3_tt._bits[0] == ~t1_cell.carry_truth_table )
-      {
-        maj_phase = true;
-      }
-      else if ( maj3_tt._bits[0] != t1_cell.carry_truth_table )
-      {
-        continue;
-      }
-
-      bool or_phase{ false };
-      if (!t1_cell.check_cbar(or3_tt, or_phase))
-      {
-        continue;
-      }
-
-      /* managed to find a T1 cell-based implementation      */
-      /* figure out the nodes connecting to the output ports */
-      klut::node carry_to     = ntk.get_constant( false );
-      klut::node inv_carry_to = ntk.get_constant( false );
-      klut::node cbar_to      = ntk.get_constant( false );
-      klut::node inv_cbar_to  = ntk.get_constant( false );
-      config_t1_ports_connection( ntk, maj_phase, maj_index, carry_to, inv_carry_to );
-      config_t1_ports_connection( ntk,  or_phase, or_index,  cbar_to,  inv_cbar_to );
-
-      /* instantiate an T1 cell */
-      t1_candidates.emplace( target_leaves, T1_OUTPUTS( 
-        t1_cell.in_phase, 
-        inv_carry_to != ntk.get_constant( false ), 
-        inv_cbar_to != ntk.get_constant( false ), 
-        false, 
-        carry_to != ntk.get_constant( false ), 
-        cbar_to != ntk.get_constant( false ), 
-        ntk.get_constant( false ), 
-        carry_to, 
-        inv_carry_to, 
-        cbar_to, 
-        inv_cbar_to ) );
-      break;
-    }
-  }
-
-  return t1_candidates;
-}
-
 template <typename Ntk>
 glob_phase_t latest_fanin_phase(const Ntk & ntk, const typename Ntk::signal & node, const uint8_t n_phases, const uint8_t type, const bool verbose)
 {
@@ -713,6 +691,591 @@ glob_phase_t latest_fanin_phase(const Ntk & ntk, const typename Ntk::signal & no
   // DEBUG_PRINT("\t{} GATE {} placed at ɸ={} [S={}, φ={}]\n",  node, phase, phase/n_phases, phase%n_phases);
 
   return phase;
+}
+
+void greedy_ntk_assign(const klut & ntk, const std::unordered_map<klut::signal, Primitive<klut>> & sig_params, const uint8_t n_phases, const std::unordered_map<unsigned int, unsigned int> & phase_assignment, const bool verbose = false)
+{
+  ( void )sig_params;
+  mockturtle::topo_view<klut> ntk_topo ( ntk );
+
+  // ntk.set_value(sig, node_data.value);
+
+  ntk_topo.foreach_node([&] ( const klut::signal & node ) 
+  {
+    if ( ntk_topo.is_constant( node ) )
+    {
+      return;
+    }
+
+    if ( ntk_topo.is_pi( node ) )
+    {
+      NodeData node_data;
+      node_data.type = AS_GATE;
+      auto ct = phase_assignment.count(node);
+      node_data.sigma = (ct != 0) ? phase_assignment.at( node ) : 0;
+      ntk.set_value(node, node_data.value);
+
+      DEBUG_PRINT("PI {} placed at ɸ=0 [S=0, φ=0]\n", node);
+      return;
+    }
+
+    // const Primitive<klut> & node_params = sig_params.at( node );
+    // if (verbose) fmt::print("{} GATE {}:\n", GATE_TYPE.at(node_params.type), node);
+
+    uint8_t node_type = NodeData( ntk.value( node ) ).type;
+    // const Primitive<klut> & prim = sig_params.at( node );
+    DEBUG_PRINT("{} GATE {}:\n", GATE_TYPE.at(node_type), node);
+
+    NodeData node_data;
+    node_data.type = node_type;
+
+    auto ct = phase_assignment.count(node);
+    if (ct != 0) // there is a precalculated phase assignment
+    {
+      node_data.sigma = phase_assignment.at( node );
+      ntk.set_value(node, node_data.value);
+    }
+    else
+    {
+      if ( node_type == AA_GATE )
+      {
+        node_data.sigma = latest_fanin_phase(ntk_topo, node, n_phases, AA_GATE, verbose);
+        uint32_t sigma = static_cast<uint32_t>(node_data.sigma);
+        DEBUG_PRINT("\tAA GATE {} placed at ɸ={} [S={}, φ={}]\n", node, sigma, sigma/n_phases, sigma%n_phases);
+      }
+      else if ( node_type == SA_GATE )
+      {
+        node_data.sigma = latest_fanin_phase(ntk_topo, node, n_phases, SA_GATE, verbose);
+        uint32_t sigma = static_cast<uint32_t>(node_data.sigma);
+        DEBUG_PRINT("\tSA GATE {} placed at ɸ={} [S={}, φ={}]\n", node, sigma, sigma/n_phases, sigma%n_phases);
+      }
+      else if ( node_type == AS_GATE )
+      {
+        node_data.sigma = latest_fanin_phase(ntk_topo, node, n_phases, SA_GATE, verbose) + 1;
+        uint32_t sigma = static_cast<uint32_t>(node_data.sigma);
+        DEBUG_PRINT("\tAS GATE {} placed at ɸ={} [S={}, φ={}]\n", node, sigma, sigma/n_phases, sigma%n_phases);
+      }
+      else 
+      {
+        std::vector<klut::signal> fanins;
+        ntk.foreach_fanin(node, [&](const klut::signal & fanin){ fanins.push_back(fanin); });
+        fmt::print("\t GATE {0} : fanins[{1}], type[{2}], func[{3}=0x{3:x}=0b{3:b}]\n", node, fmt::join(fanins, ","), node_type, ntk.node_function(node)._bits[0]);
+        throw "Unsupported gate type!";
+      }
+    }
+  });
+}
+
+
+std::unordered_map<klut::signal, glob_phase_t> greedy_assign(const klut & ntk, const std::unordered_map<klut::signal, Primitive<klut>> & sig_params,  const bool verbose = false) //const uint8_t n_phases,
+{
+  mockturtle::topo_view<klut> ntk_topo ( ntk );
+  std::unordered_map<klut::signal, glob_phase_t> glob_phase;
+  glob_phase.reserve(ntk_topo.size());
+
+  printUnorderedMap(sig_params);
+
+  ntk_topo.foreach_node([&] ( const klut::signal & node ) 
+  {
+    if ( ntk_topo.is_constant( node ) )
+    {
+      DEBUG_PRINT("Skipping constant {}\n", node);
+      return;
+    }
+    if ( ntk_topo.is_pi( node ) )
+    {
+      glob_phase[node] = 0u;
+      DEBUG_PRINT("PI {} placed at ɸ=0 [S=0, φ=0]\n", node);
+      return;
+    }
+    const Primitive<klut> & node_params = sig_params.at( node );
+    DEBUG_PRINT("{} GATE {}:\n", GATE_TYPE.at(node_params.type), node);
+
+    if ( node_params.type == AA_GATE )
+    {
+      // place at the same phase as the latest fanin
+      glob_phase_t phase = 0u;
+      ntk_topo.foreach_fanin(node, [&] ( const klut::signal & parent )
+      {
+        if ( ntk_topo.is_constant( parent ) )
+        {
+        DEBUG_PRINT("\t\tfanin {} is constant, skipping...\n", parent);
+        return;
+        }
+        phase = std::max(phase, glob_phase.at( parent ));
+        DEBUG_PRINT("\t\tfanin {} ɸ={}\n", parent, glob_phase.at( parent ));
+        // fmt::print("\t\tfanin {} ɸ={} [S={}, φ={}]\n", parent, gp, gp/n_phases, gp%n_phases);
+      });
+      glob_phase[node] = phase;
+    if (verbose)
+    {
+      fmt::print("\tAA GATE {} placed at ɸ={}\n", node, phase);
+      // fmt::print("\tAA GATE {} placed at ɸ={} [S={}, φ={}]\n", node, phase, phase/n_phases, phase%n_phases);
+    }
+    }
+    else if ( node_params.type == SA_GATE )
+    {
+    // Place at the earliest feasible phase
+    glob_phase_t phase = 0u;
+    ntk_topo.foreach_fanin(node, [&] (const klut::signal & parent)
+    {
+        if ( ntk_topo.is_constant( parent ) )
+        {
+        if (verbose)
+        {
+            fmt::print("\t\tfanin {} is constant, skipping...\n", parent);
+        }
+        return;
+        }
+        // if d==false, SA gate can be directly connected to fanin
+        //cannot directly connect PI
+        //cannot directly connect split signal 
+        //cannot directly connect SA/AA gates to SA gates
+        bool d = (ntk_topo.is_pi(parent)) || (ntk_topo.fanout_size(parent) > 1) || (sig_params.at(parent).type != AS_GATE);                     
+
+        phase = std::max(phase, glob_phase.at( parent ) + static_cast<int>(d) );
+        if (verbose)
+        {
+        glob_phase_t gp = glob_phase.at( parent );
+        std::string gt;
+        // fmt::print("\t\tfanin: {4} {0} ɸ={1} [S={2}, φ={3}], fanout={5} \n",
+        fmt::print("\t\tfanin: {2} {0} ɸ={1}, fanout={3} \n", 
+            parent, 
+            gp, 
+            // gp/n_phases, 
+            // gp%n_phases, 
+            ntk_topo.is_pi(parent) ? "PI" : GATE_TYPE.at( sig_params.at(parent).type ),
+            ntk_topo.fanout_size(parent)
+        );
+        }
+    });
+    glob_phase[node] = phase;
+    if (verbose)
+    {
+        // fmt::print("\tSA GATE {} placed at ɸ={} [S={}, φ={}]\n", node, phase, phase/n_phases, phase%n_phases);
+        fmt::print("\tSA GATE {} placed at ɸ={}\n", node, phase);
+    }
+    }
+    else if ( node_params.type == AS_GATE )
+    {
+    // Place at the earliest feasible phase
+    glob_phase_t phase = 0u;
+    ntk_topo.foreach_fanin(node, [&] (const klut::signal & parent)
+    {
+        if ( ntk_topo.is_constant( parent ) )
+        {
+        if (verbose)
+        {
+            fmt::print("\t\tfanin {} is constant, skipping...\n", parent);
+        }
+        return;
+        }
+        phase = std::max(phase, glob_phase.at( parent ) );
+        if (verbose)
+        {
+        glob_phase_t gp = glob_phase.at( parent );
+        // fmt::print("\t\tfanin {} ɸ={} [S={}, φ={}]\n", parent, gp, gp/n_phases, gp%n_phases);
+        fmt::print("\t\tfanin {} ɸ={}\n", parent, gp);
+        }
+    });
+    phase++;
+    glob_phase[node] = phase;
+    if (verbose)
+    {
+        fmt::print("\tAS GATE {} placed at ɸ={}\n", node, phase);
+        // fmt::print("\tAS GATE {} placed at ɸ={} [S={}, φ={}]\n", node, phase, phase/n_phases, phase%n_phases);
+    }
+    }
+    else 
+    {
+    fmt::print("\t GATE {} : fanins[{}], type[{}], func[{}]\n", node, fmt::join(node_params.fanins, ","), node_params.type, node_params.func);
+    throw "Unsupported gate type!";
+    }
+  });
+  return glob_phase;
+}
+
+bool phase_ntk_comparison(const klut::signal & a, const klut::signal & b, const klut & ntk )
+{
+  if (ntk.is_constant(a))
+  {
+    return true;
+  }
+  else if (ntk.is_constant(b))
+  {
+    return false;
+  }
+  NodeData a_data = ntk.value(a);
+  NodeData b_data = ntk.value(b);
+
+  return a_data.sigma < b_data.sigma;
+}
+
+// Function to insert splitter nodes in a KLUT-based network (klut = k-input lookup table network).
+void splitter_ntk_insertion(klut & ntk, const bool verbose = false)
+{
+  // Lambda function for comparing the phases of two signals in the network.
+  auto phase_comp = [&](const klut::signal & a, const klut::signal & b)
+  {
+    return phase_ntk_comparison(a, b, ntk);
+  };
+
+  // Create a view of the network that provides access to fanout information.
+  auto ntk_fo = mockturtle::fanout_view<klut>(ntk);
+
+  // For each node in the fanout view:
+  ntk_fo.foreach_node([&](const klut::signal & node)
+  {
+    // Get the number of fanouts for the current node.
+    uint32_t fo_size = ntk_fo.fanout_size(node);
+
+    DEBUG_PRINT("\t[NODE {}] FANOUT SIZE = {}\n", node, fo_size);
+    // If the current node is a constant or it has fanout ≤ 1, skip to the next node.
+    if (ntk_fo.is_constant(node) || fo_size <= 1)
+    {
+      return;
+    }
+
+    // Populate the fanouts vector.
+    std::vector<klut::signal> fanouts;
+    fanouts.reserve(fo_size);
+    ntk_fo.foreach_fanout(node, [&](const klut::signal & fo_node)
+    {
+      fanouts.push_back(fo_node);
+      DEBUG_PRINT("\t\t[NODE {}] ADDING FANOUT\n", node, fo_node);
+    });
+
+    // Fix the fanout count (bugged fanouts_size()?)
+    if ( fanouts.size() != fo_size )
+    {
+      ntk._storage->nodes[node].data[0].h1 = fanouts.size();
+    }
+
+    // Sort the fanouts using the phase comparison function.
+    std::sort(fanouts.begin(), fanouts.end(), phase_comp);
+    DEBUG_PRINT("\t[NODE {}] SORTED FANOUTS:\n", node);
+    printVector(fanouts, 2);
+
+    // Create [fo_size - 1] splitter nodes.
+    klut::signal last_spl = node;
+    std::vector<klut::signal> splitters;
+    splitters.reserve( fanouts.size() - 1 );
+    for (auto it = fanouts.begin(); it < fanouts.end() - 1; it++)
+    {
+      DEBUG_PRINT("\t\t[NODE {}] LAST SPL: {}\n", node, last_spl);
+      // Copy sigma and type data from the first fanout for the splitter node data.
+      NodeData spl_data = ntk.value(*it);
+      // Change gate type to AA for the splitter node.
+      spl_data.type = AA_GATE;
+      // Create a new splitter node connected to 'last_spl'.
+
+      DEBUG_PRINT("\t\t[NODE {}] CREATING SPL FOR {}\n", node, *it);
+      DEBUG_PRINT("\t\t[NODE {}] LAST_SPL {} FANOUT BEFORE: {}\n", node, last_spl, ntk.fanout_size(last_spl));
+      const klut::signal spl = ntk._create_node({last_spl}, 2, last_spl);
+      ntk.set_value(spl, spl_data.value);
+      DEBUG_PRINT("\t\t[NODE {}] CREATED SPL {}\n", node, spl);
+      DEBUG_PRINT("\t\t[NODE {}] LAST_SPL {} FANOUT AFTER: {}\n", node, last_spl, ntk.fanout_size(last_spl));
+      DEBUG_PRINT("\t\t[NODE {}] SPL {} FANIN: {}\n", node, spl, ntk.fanin_size(spl));
+
+      // auto n = ntk._storage->nodes[*it];
+      // auto& preds = n.children;
+
+      // Update the connections to reflect the splitter's presence.
+      DEBUG_PRINT("\t\t[NODE {}, LAST_SPL {}] UPDATING CONNECTIONS\n", node, last_spl);
+      // for (auto& pred : preds)
+      for (auto pred_it = ntk._storage->nodes[*it].children.begin();
+                pred_it < ntk._storage->nodes[*it].children.end(); pred_it++ )
+      {
+        DEBUG_PRINT("\t\t\t[NODE {}, LAST_SPL {}] PRED {}\n", node, last_spl, pred_it->data);
+        if (pred_it->data == node)
+        {
+          DEBUG_PRINT("\t\t\t[NODE {}, LAST_SPL {}] RECORDING OLD PREDS\n", node, last_spl);
+          // Store the previous connections.
+          std::vector<klut::signal> old_preds(ntk._storage->nodes[*it].children.size());
+          std::transform(ntk._storage->nodes[*it].children.begin(), ntk._storage->nodes[*it].children.end(), old_preds.begin(), [](auto c) { return c.index; });
+          printVector(old_preds, 4);
+
+          DEBUG_PRINT("\t\t\t[NODE {}, LAST_SPL {}] PREDS BEFORE\n", node, last_spl);
+          for (const auto& entry : ntk._storage->nodes[*it].children)  { fmt::print("\t\t\t\t{}\n", entry.data); }
+          pred_it->data = spl;                             // Replace the connection with the splitter.
+          DEBUG_PRINT("\t\t\t[NODE {}, LAST_SPL {}] PREDS AFTER\n", node, last_spl);
+          for (const auto& entry : ntk._storage->nodes[*it].children)  { fmt::print("\t\t\t\t{}\n", entry.data); }
+
+          DEBUG_PRINT("\t\t\t[NODE {}, LAST_SPL {}] SPL {} FANOUT BEFORE: {}\n", node, last_spl, spl, static_cast<int>(ntk._storage->nodes[spl].data[0].h1));
+          DEBUG_PRINT("\t\t\t\t Call: {}\n", ntk.fanout_size(spl));
+          ntk._storage->nodes[spl].data[0].h1++;  // Increment fan-out of the splitter.
+          DEBUG_PRINT("\t\t\t[NODE {}, LAST_SPL {}] SPL {} FANOUT  AFTER: {}\n", node, last_spl, spl, static_cast<int>(ntk._storage->nodes[spl].data[0].h1));
+          DEBUG_PRINT("\t\t\t\t Call: {}\n", ntk.fanout_size(spl));
+
+          DEBUG_PRINT("\t\t\t[NODE {}, LAST_SPL {}] NODE {} FANOUT BEFORE: {}\n", node, last_spl, node, static_cast<int>(ntk._storage->nodes[node].data[0].h1));
+          DEBUG_PRINT("\t\t\t\t Call: {}\n", ntk.fanout_size(node));
+          ntk._storage->nodes[node].data[0].h1--; // Decrement fan-out of the current node.
+          DEBUG_PRINT("\t\t\t[NODE {}, LAST_SPL {}] NODE {} FANOUT  AFTER: {}\n", node, last_spl, node, static_cast<int>(ntk._storage->nodes[node].data[0].h1));
+          DEBUG_PRINT("\t\t\t\t Call: {}\n", ntk.fanout_size(node));
+
+          // // Notify listeners of the modification.
+          // for (auto const& fn : ntk._events->on_modified)
+          // {
+          //   (*fn)(*it, old_preds);
+          // }
+        }
+      }
+      last_spl = spl;
+      splitters.push_back(spl);
+    }
+
+    // Process the last fanout.
+    // auto& preds = ntk._storage->nodes[fanouts.back()].children;
+    DEBUG_PRINT("\t\t[NODE {}, LAST_SPL {}] UPDATING CONNECTIONS TO LAST FANOUT {}\n", node, last_spl, fanouts.back());
+    // for (auto& pred : preds)
+    for (auto pred_it = ntk._storage->nodes[fanouts.back()].children.begin();
+              pred_it < ntk._storage->nodes[fanouts.back()].children.end(); pred_it++ )
+    {
+      if (pred_it->data == node)
+      {
+        DEBUG_PRINT("\t\t\t[NODE {}, LAST_SPL {}] RECORDING OLD PREDS\n", node, last_spl);
+        // Store the previous connections.
+        std::vector<klut::signal> old_preds(ntk._storage->nodes[fanouts.back()].children.size());
+        std::transform(ntk._storage->nodes[fanouts.back()].children.begin(), ntk._storage->nodes[fanouts.back()].children.end(), old_preds.begin(), [](auto c) { return c.index; });
+        printVector(old_preds, 4);
+
+        DEBUG_PRINT("\t\t\t[NODE {}, LAST_SPL {}] PREDS BEFORE\n", node, last_spl);
+        for (const auto& entry : ntk._storage->nodes[fanouts.back()].children)  { fmt::print("\t\t\t\t{}\n", entry.data); }
+        pred_it->data = last_spl;                            // Replace the connection with the last splitter.
+        DEBUG_PRINT("\t\t\t[NODE {}, LAST_SPL {}] PREDS AFTER\n", node, last_spl);
+        for (const auto& entry : ntk._storage->nodes[fanouts.back()].children)  { fmt::print("\t\t\t\t{}\n", entry.data); }     
+
+
+        DEBUG_PRINT("\t\t\t[NODE {}, LAST_SPL {}] SPL {} FANOUT BEFORE: {}\n", node, last_spl, last_spl, static_cast<int>(ntk._storage->nodes[last_spl].data[0].h1));
+        DEBUG_PRINT("\t\t\t\t Call: {}\n", ntk.fanout_size(last_spl));
+        ntk._storage->nodes[last_spl].data[0].h1++; // Increment fan-out of the last splitter.
+        DEBUG_PRINT("\t\t\t[NODE {}, LAST_SPL {}] SPL {} FANOUT  AFTER: {}\n", node, last_spl, last_spl, static_cast<int>(ntk._storage->nodes[last_spl].data[0].h1));
+        DEBUG_PRINT("\t\t\t\t Call: {}\n", ntk.fanout_size(last_spl));
+
+        DEBUG_PRINT("\t\t\t[NODE {}, LAST_SPL {}] NODE {} FANOUT BEFORE: {}\n", node, last_spl, node, static_cast<int>(ntk._storage->nodes[node].data[0].h1));
+        DEBUG_PRINT("\t\t\t\t Call: {}\n", ntk.fanout_size(node));
+        ntk._storage->nodes[node].data[0].h1--;     // Decrement fan-out of the current node.
+        DEBUG_PRINT("\t\t\t[NODE {}, LAST_SPL {}] NODE {} FANOUT  AFTER: {}\n", node, last_spl, node, static_cast<int>(ntk._storage->nodes[node].data[0].h1));
+        DEBUG_PRINT("\t\t\t\t Call: {}\n", ntk.fanout_size(node));
+
+        // Notify listeners of the modification.
+        for (auto const& fn : ntk._events->on_modified)
+        {
+          (*fn)(fanouts.back(), old_preds);
+        }
+      }
+    }
+
+    auto fo_ntk { mockturtle::fanout_view( ntk ) };
+    for (const auto & node : fanouts)
+    {
+      DEBUG_PRINT("\t\t\t node: {}\n", node);
+      
+      fo_ntk.foreach_fanin( node, [&](const klut::signal & fi_node)
+      {
+        DEBUG_PRINT("\t\t\t\t fanin : {}\n", fi_node);
+      });
+      fo_ntk.foreach_fanout( node, [&](const klut::signal & fo_node)
+      {
+        DEBUG_PRINT("\t\t\t\t fanout: {}\n", fo_node);
+      });
+    }
+    for (const auto & node : splitters)
+    {
+      DEBUG_PRINT("\t\t\t spl : {}\n", node);
+      
+      fo_ntk.foreach_fanin( node, [&](const klut::signal & fi_node)
+      {
+        DEBUG_PRINT("\t\t\t\t fanin : {}\n", fi_node);
+      });
+      fo_ntk.foreach_fanout( node, [&](const klut::signal & fo_node)
+      {
+        DEBUG_PRINT("\t\t\t\t fanout: {}\n", fo_node);
+      });
+    }
+    // Ensure that the current node's fan-out count is now 1 (since all other fanouts have been replaced by splitters).
+    assert(ntk._storage->nodes[node].data[0].h1 == 1);
+  });
+}
+
+uint64_t calculate_hash(glob_phase_t phase, klut::signal fanout, klut::signal fanin)
+{
+    std::size_t seed = 0;
+    hash_combine(seed, phase);
+    hash_combine(seed, fanout);
+    hash_combine(seed, fanin);
+    return seed;
+};
+
+struct DFF_var 
+{
+  klut::signal fanin;
+  klut::signal fanout;
+  uint32_t sigma;
+  std::unordered_set<uint64_t> parent_hashes;
+  uint64_t hash;
+
+  DFF_var(klut::signal _fanin, klut::signal _fanout, uint32_t _sigma)
+      : fanin(_fanin), fanout(_fanout), sigma(_sigma), parent_hashes({}) 
+      {
+        hash = calculate_hash(_sigma, _fanout, _fanin);
+      }
+
+  DFF_var(klut::signal _fanin, klut::signal _fanout, uint32_t _sigma, std::unordered_set<uint64_t> _parent_hashes, uint64_t _hash)
+      : fanin(_fanin), fanout(_fanout), sigma(_sigma), parent_hashes(_parent_hashes), hash(_hash) {}
+
+  DFF_var(const DFF_var& other)
+      : fanin(other.fanin), fanout(other.fanout), sigma(other.sigma), parent_hashes(other.parent_hashes), hash(other.hash) {}
+
+  std::string str() 
+  {
+      return fmt::format("var_{}_{}_{}", fanin, fanout, sigma);
+  }
+};
+
+uint64_t dff_hash(klut::signal _fanin, klut::signal _fanout, uint32_t _sigma)
+{
+  return ( (uint64_t)_fanin << 40 ) | ( (uint64_t)_fanout << 16 ) | _sigma;
+}
+
+struct DFF_registry
+{
+  std::unordered_map<uint64_t, DFF_var> variables;
+
+  DFF_var & at(node_t _fanin, node_t _fanout, glob_phase_t _sigma)
+  {
+    return variables.at( dff_hash(_fanin, _fanout, _sigma) );
+  } 
+  DFF_var & at(uint64_t _hash)
+  {
+    return variables.at( _hash );
+  } 
+  uint64_t add(node_t _fanin, node_t _fanout, glob_phase_t _phase, std::unordered_set<uint64_t> _parent_hashes = {})
+  {
+    uint64_t _hash = dff_hash(_phase, _fanout, _fanin);
+    DFF_var temp { _fanin, _fanout, _phase, _parent_hashes, _hash };
+    variables.emplace(_hash, temp);
+    // variables.insert(std::make_pair(_hash, temp));
+    return _hash;
+  } 
+
+  std::string str(uint64_t hash, bool negated = false)
+  {
+    DFF_var & dff = variables.at(hash);
+    if (negated)
+    {
+      return fmt::format( "var_{}_{}_{}.Not()", dff.fanin, dff.fanout, dff.sigma );
+    }
+    return fmt::format( "var_{}_{}_{}", dff.fanin, dff.fanout, dff.sigma );
+  }
+};
+
+
+std::vector<Path> extract_paths(const klut & ntk, bool verbose = false)
+{
+  DEBUG_PRINT("\t[i] ENTERED FUNCTION extract_paths\n");
+  std::vector<Path> paths;
+
+  ntk.foreach_node([&](const klut::signal & fo_node)
+  {
+    DEBUG_PRINT("\t\t[i] PROCESSING NODE {}\n", fo_node);
+    if (ntk.is_constant(fo_node) || ntk.is_pi(fo_node))
+    {
+      DEBUG_PRINT("\t\t\t[NODE {}] the node is IS CONSTANT\n", fo_node);
+      return;
+    }
+    NodeData fo_node_data = ntk.value(fo_node);
+    if (fo_node_data.type != AS_GATE && fo_node_data.type != SA_GATE) 
+    {
+      DEBUG_PRINT("\t\t\t[NODE {}] the node is AA, skipping\n", fo_node);
+      return;
+    }
+    // at this point, the node should be AS/SA
+    DEBUG_PRINT("\t\t[NODE {}] the node is AS/SA, continuing...\n", fo_node);
+
+    // Create a separate path for each fanin of the node
+    std::vector<Path> aa_paths; 
+    aa_paths.reserve( ntk.fanin_size(fo_node) );
+
+    ntk.foreach_fanin(fo_node, [&](const klut::signal & fi_node)
+    {
+      DEBUG_PRINT("\t\t\t[NODE {}] processing fanin {}\n", fo_node, fi_node);
+      // Create a path object with only a target
+      Path node_path;
+      node_path.targets.emplace( fo_node );
+
+      std::vector<klut::signal> stack { fi_node };
+      
+      DEBUG_PRINT("\t\t\t[NODE {}][FANIN {}] created stack\n", fo_node, fi_node);
+      
+      std::set<klut::signal> seen;
+      while (!stack.empty())
+      {
+        DEBUG_PRINT("\t\t\t[NODE {}][FANIN {}] stack contents:\n", fo_node, fi_node);
+        printVector(stack, 4);
+
+        const klut::signal & n = stack.back();
+        stack.pop_back();
+
+        DEBUG_PRINT("\t\t\t[NODE {}][FANIN {}]: Analyzing node {}\n", fo_node, fi_node, n);
+
+        // A constant does not have any effect on the DFF placement, we can skip it
+        if ( ntk.is_constant( n ) )
+        {
+          continue;
+        }        
+        const NodeData n_data { ntk.value(n) };
+        // Found a source of the path, add to sources, do not continue traversal
+        if ( ntk.is_pi(n) || n_data.type == AS_GATE || n_data.type == SA_GATE )
+        {
+          DEBUG_PRINT("\t\t\t[NODE {}][FANIN {}]: node {} is a source \n", fo_node, fi_node, n);
+          node_path.sources.emplace( n );
+        }
+        // Found AA gate, add to internal nodes, add parents for further traversal
+        else if ( n_data.type == AA_GATE )
+        {
+          DEBUG_PRINT("\t\t\t[NODE {}][FANIN {}]: node is INTERNAL adding fanins \n", fo_node, fi_node, n);
+          node_path.internals.emplace( n );
+
+          ntk.foreach_fanin(n, [&](const klut::signal & sig){
+            stack.push_back( sig );
+          });
+        }
+        else
+        {
+          DEBUG_PRINT("\t\t\t[NODE {}][FANIN {}]: Signal {}: {} is not recognized \n", fo_node, fi_node, n, GATE_TYPE.at( n_data.type ));
+          throw "Unsupported case";
+        }
+        seen.emplace( n );
+      }
+
+      // Identify overlapping paths
+      std::vector<size_t> to_merge;
+      for (size_t i = 0u; i < paths.size(); ++i)
+      {
+        Path & known_paths = paths[i];
+        // merge if there are sources in common
+        if( haveCommonElements( known_paths.sources, node_path.sources) )
+        {
+          to_merge.push_back(i);
+        }
+      }
+      // Merge overlapping paths into the path object and remove the path object
+      // Iterating in reverse order to preserve order
+      for (auto it = to_merge.rbegin(); it != to_merge.rend(); ++it) 
+      {
+        auto idx = *it;
+        // fmt::print("Before absorption\n");
+        // node_path.print();
+        node_path.absorb(paths[idx]);
+        // fmt::print("After absorption\n");
+        // node_path.print();
+        paths.erase(paths.begin() + idx);
+      }
+      paths.push_back(node_path);
+    });
+  });
+  return paths;
 }
 
 /// @brief Create binary variables for DFF placement in a given path
@@ -962,6 +1525,32 @@ std::vector<Snake> sectional_snake(const Path & path, klut & ntk,  DFF_registry 
   return out_snakes;
 }
 
+void write_klut_specs(const klut & ntk, const std::unordered_map<klut::signal, Primitive<klut>> & sig_params, const std::unordered_map<klut::signal, glob_phase_t> &glob_phase, const std::string & filename)
+{
+  auto ntk_fo = mockturtle::fanout_view<klut, false>( ntk );
+  std::ofstream spec_file (filename);
+  // ntk_fo.foreach_pi([&] (const auto & sig) 
+  // {
+  //   std::vector<klut::node> fanouts;
+  //   ntk_fo.foreach_fanout(sig, [&](const klut::node fo)
+  //   {
+  //     fanouts.push_back(fo);
+  //   });
+  //   spec_file << fmt::format("PI {0},{1}\n", sig, fmt::join(fanouts, "|"));
+  // } );
+  for (const auto & [sig, prim] : sig_params)
+  {
+    std::vector<klut::node> fanouts;
+    ntk_fo.foreach_fanout(sig, [&](const klut::node fo)
+    {
+      fanouts.push_back(fo);
+    });
+
+    uint8_t prim_type = prim.fanins.empty() ? PI_GATE : prim.type;
+    spec_file << fmt::format("{0},{1},{2},{3},{4},{5}\n", sig, prim.func, prim_type, fmt::join(prim.fanins, "|"), glob_phase.at(sig), fmt::join(fanouts, "|"));
+  }
+}
+
 std::tuple<int, std::unordered_map<unsigned int, unsigned int>, std::string>  cpsat_macro_opt(const std::string & cfg_name, uint8_t n_phases) 
 {
   std::string command = fmt::format("{} {} {} {}", PYTHON_EXECUTABLE, PYTHON_PHASE_ASSIGNMENT, n_phases, cfg_name);
@@ -1124,11 +1713,13 @@ int cpsat_ortools(const std::string & cfg_name)
   }
 }
 
-void config_t1_ports_connection( klut const& ntk, const bool output_phase, const uint32_t target_node, klut::node& port, klut::node& inv_port )
+void config_t1_ports_connection( klut const& ntk, const bool output_phase, 
+                                 const uint32_t target_node, 
+                                 klut::node& port, klut::node& inv_port )
 {
   if ( !output_phase )
   {
-    /*      connect the node to the output port      */
+    /* connect the node to the output port */
     port = ntk.index_to_node( target_node );
     /* if (1) the target node is an inverter, and    */
     /* (2) its fanin node has more than one fanout,  */
@@ -1159,6 +1750,43 @@ void config_t1_ports_connection( klut const& ntk, const bool output_phase, const
   }
 }
 
+/* this function is merely for debugging and can be removed */
+std::string get_gate_type( const uint32_t gate_type )
+{
+  switch( gate_type )
+  {
+  case 3:
+    return "a NOT gate";
+  case 4:
+    return "an AND gate";
+  case 6:
+    return "an OR gate";
+  case 12:
+    return "an XOR gate";
+  default:
+    return "an unknown gate with the functional literal of " + std::to_string( gate_type );
+  }
+}
+
+/* this function is merely for debugging and can be removed */
+void profile_klut_node( klut const& ntk, klut::node const& n )
+{
+  mockturtle::fanout_view<klut, false> ntk_fanout{ ntk };
+  fmt::print( "Profiling Node {} :\n", n );
+  fmt::print( "\tGate type : {}\n", get_gate_type( ntk.func_lit( n ) ) );
+  fmt::print( "\tFanins : " );
+  /* notice that 'signal' and 'node' are the same concept in KLUTs */
+  ntk.foreach_fanin( n, []( auto const& ni ) {
+    fmt::print( "Node {}  ", ni );
+  } );
+  fmt::print( "\n" );
+  fmt::print( "\tFanouts : " );
+  ntk_fanout.foreach_fanout( n, []( auto const& no ) {
+    fmt::print( "Node {}  ", no );
+  } );
+  fmt::print( "\n" );
+}
+
 uint32_t get_node_cost( const uint32_t gate_type )
 {
   switch( gate_type )
@@ -1179,17 +1807,18 @@ uint32_t get_node_cost( const uint32_t gate_type )
 
 uint32_t deref_node( klut& ntk, const std::array<uint32_t, 3> leaves, const uint32_t n )
 {
-  /* return the number of nodes in the original cut that can be removed, */
-  /* if the function of current cut is implemented using an T1 cell,     */
-  /* i.e., the MFFC size                                                 */
-  if ( auto it_find = std::find( leaves.begin(), leaves.end(), n ); it_find != leaves.end() )
+  /* return the total cost of nodes in the original cut that can be      */
+  /* removed, if implemented using an T1 cell                            */
+  if ( std::find( leaves.begin(), leaves.end(), n ) != leaves.end() )
   {
     return 0u;
   }
 
   uint32_t gain = get_node_cost( ntk.func_lit( n ) );
   ntk.foreach_fanin( n, [&]( auto const& ni ) {
-    if ( ntk.decr_fanout_size( ni ) == 0 )
+    /* skip fanins that are leaves */
+    if ( ( std::find( leaves.begin(), leaves.end(), ni ) == leaves.end() ) && 
+         ( ntk.decr_fanout_size( ni ) == 0 ) )
     {
       gain += deref_node( ntk, leaves, ntk.node_to_index( ni ) );
     }
@@ -1214,7 +1843,11 @@ void reref_node( klut& ntk, const std::array<uint32_t, 3> leaves, const uint32_t
   }
 
   ntk.foreach_fanin( n, [&]( auto const& ni ) {
-    reref_node( ntk, leaves, ntk.node_to_index( ni ) );
+    if ( ( std::find( leaves.begin(), leaves.end(), ni ) == leaves.end() ) && 
+         ( ntk.incr_fanout_size( ni ) == 0 ) )
+    {
+      reref_node( ntk, leaves, ntk.node_to_index( ni ) );
+    }
   } );
 }
 
@@ -1226,7 +1859,7 @@ bool t1_usage_sanity_check( klut& ntk, std::pair<const std::array<uint32_t, 3>, 
   std::array<uint32_t, 3> roots;
   roots[0] = ntk.node_to_index( t1_outputs.sum_to );
   roots[1] = std::max( ntk.node_to_index( t1_outputs.carry_to ), ntk.node_to_index( t1_outputs.inv_carry_to ) );
-  roots[2] = std::max( ntk.node_to_index( t1_outputs.cbar_to  ), ntk.node_to_index( t1_outputs.inv_cbar_to  ) );
+  roots[2] = std::max( ntk.node_to_index(  t1_outputs.cbar_to ), ntk.node_to_index(  t1_outputs.inv_cbar_to ) );
 
   for ( const uint32_t root : roots )
   {
@@ -1246,19 +1879,20 @@ bool t1_usage_sanity_check( klut& ntk, std::pair<const std::array<uint32_t, 3>, 
 
     return false;
   }
+
   /* update gate type for the committed T1 cells */
-  if ( t1_outputs.has_sum ) { ntk.set_value( t1_outputs.sum_to, NodeData( static_cast<NodeData>( ntk.value( t1_outputs.sum_to ) ).sigma, T1_GATE ).value ); }
-  if ( t1_outputs.has_carry ) { ntk.set_value( t1_outputs.carry_to, NodeData( static_cast<NodeData>( ntk.value( t1_outputs.carry_to ) ).sigma, T1_GATE ).value ); }
-  if ( t1_outputs.has_carry_inverted ) { ntk.set_value( t1_outputs.inv_carry_to, NodeData( static_cast<NodeData>( ntk.value( t1_outputs.inv_carry_to ) ).sigma, T1_GATE ).value ); }
-  if ( t1_outputs.has_cbar ) { ntk.set_value( t1_outputs.cbar_to, NodeData( static_cast<NodeData>( ntk.value( t1_outputs.cbar_to ) ).sigma, T1_GATE ).value ); }
-  if ( t1_outputs.has_cbar_inverted ) { ntk.set_value( t1_outputs.inv_cbar_to, NodeData( static_cast<NodeData>( ntk.value( t1_outputs.inv_cbar_to ) ).sigma, T1_GATE ).value ); }
+  if ( t1_outputs.has_sum ) { ntk.set_value( t1_outputs.sum_to, NodeData( static_cast<NodeData>( ntk.value( t1_outputs.sum_to ) ).sigma, FA_GATE ).value ); }
+  if ( t1_outputs.has_carry ) { ntk.set_value( t1_outputs.carry_to, NodeData( static_cast<NodeData>( ntk.value( t1_outputs.carry_to ) ).sigma, FA_GATE ).value ); }
+  if ( t1_outputs.has_carry_inverted ) { ntk.set_value( t1_outputs.inv_carry_to, NodeData( static_cast<NodeData>( ntk.value( t1_outputs.inv_carry_to ) ).sigma, FA_GATE ).value ); }
+  if ( t1_outputs.has_cbar ) { ntk.set_value( t1_outputs.cbar_to, NodeData( static_cast<NodeData>( ntk.value( t1_outputs.cbar_to ) ).sigma, FA_GATE ).value ); }
+  if ( t1_outputs.has_cbar_inverted ) { ntk.set_value( t1_outputs.inv_cbar_to, NodeData( static_cast<NodeData>( ntk.value( t1_outputs.inv_cbar_to ) ).sigma, FA_GATE ).value ); }
 
   updated_area -= gain;
 
   return true;
 }
 
-void write_klut_specs_supporting_t1( klut const& ntk, array_map<3, T1_OUTPUTS> const& t1s, std::string const& filename )
+void write_klut_specs_supporting_t1( klut const& ntk, phmap::flat_hash_map<std::array<uint32_t,3>, T1_OUTPUTS, ArrayHash<3>> const& t1s, std::string const& filename )
 {
   std::ofstream spec_file( filename );
 
@@ -1286,8 +1920,8 @@ void write_klut_specs_supporting_t1( klut const& ntk, array_map<3, T1_OUTPUTS> c
     return true;
   } );
 
-  /* write information of the committed T1 cells into  */
-  /* the csv file                                      */
+  /* write information of the commited T1 cells into  */
+  /* the csv file                                     */
   for ( auto const& t1 : t1s )
   {
     /* the 5 output ports are in the order of: sum,   */
@@ -1413,8 +2047,8 @@ int main(int argc, char* argv[])  //
 
 #pragma region benchmark_parsing
   // *** BENCHMARKS OF INTEREST ***
-  auto benchmarks1 = epfl_benchmarks( experiments::arithmetic );
-  //auto benchmarks1 = epfl_benchmarks( experiments::adder );
+  //auto benchmarks1 = epfl_benchmarks( experiments::arithmetic );
+  auto benchmarks1 = epfl_benchmarks( experiments::adder );
   //auto benchmarks2 = iscas_benchmarks( experiments::c432 | experiments::c880 | experiments::c1908 | experiments::c1355 | experiments::c3540 );
   //benchmarks1.insert(benchmarks1.end(), benchmarks2.begin(), benchmarks2.end());
 
@@ -1496,15 +2130,16 @@ int main(int argc, char* argv[])  //
 
 
     // *** DECOMPOSE COMPOUND GATES INTO PRIMITIVES, REMOVE DFFS, REPLACE OR GATES WITH CB WHERE POSSIBLE ***
-    auto _result = decompose_to_klut(res_wo_pb, GNM_global, entries, COSTS_MAP);
+    auto _result = decompose_to_klut(res_wo_pb, GNM_global, entries);
     auto klut_decomposed = std::get<0>(_result);
-    auto raw_area = std::get<1>(_result);
+    auto klut_prim_params = std::get<1>(_result);
+    auto raw_area = std::get<2>(_result);
     fmt::print("Decomposition complete\n");
 
-    // *** ENUMERATE 3-CUTS ***
     cut_enumeration_params ce_params; 
     ce_params.cut_size = 3u;
-    const auto cuts = mockturtle::cut_enumeration<klut, true>( klut_decomposed, ce_params );
+  
+    auto cuts = mockturtle::cut_enumeration<klut, true>( klut_decomposed, ce_params );
 
     /* print enumerated cuts */
     // klut_decomposed.foreach_node( [&]( auto node ) {
@@ -1513,33 +2148,279 @@ int main(int argc, char* argv[])  //
     //   std::cout << node_cuts << "\n";
     // } );
 
-    // *** FIND THOSE CUTS MATCHING THE XOR3/MAJ3/OR3 FUNCTIONS ***
-    const auto xor3_cuts = match_cuts( xor3_tts, klut_decomposed, cuts);
-    const auto maj3_cuts = match_cuts( maj3_tts, klut_decomposed, cuts);
-    const auto  or3_cuts = match_cuts(  or3_tts, klut_decomposed, cuts);
+    phmap::flat_hash_map<std::array<uint32_t,3>, std::tuple<kitty::dynamic_truth_table, uint32_t>, ArrayHash<3>> maj3_cuts;
+    phmap::flat_hash_map<std::array<uint32_t,3>, std::tuple<kitty::dynamic_truth_table, uint32_t>, ArrayHash<3>> xor3_cuts;
+    phmap::flat_hash_map<std::array<uint32_t,3>, std::tuple<kitty::dynamic_truth_table, uint32_t>, ArrayHash<3>>  or3_cuts;
+    // std::unordered_set<std::array<uint32_t,4>> maj3_cuts;
+    // std::unordered_set<std::array<uint32_t,4>> xor3_cuts;
+    // phmap::flat_hash_set<cut_type<true, empty_cut_data>> maj3_cuts;
+    // phmap::flat_hash_set<cut_type<true, empty_cut_data>> xor3_cuts;
+    klut_decomposed.foreach_node( [&]( const klut::signal & node ) { // 
+      const auto idx = klut_decomposed.node_to_index( node );
+      const auto & node_cut_set = cuts.cuts( idx );
+      // fmt::print("[{}] Extracted {} cuts for node {} (idx={})\n", benchmark, node_cut_set.size(), node, idx);
+      // if (node_cut_set.size() == 0) { return; }
+
+      for (const auto & cut_entry : node_cut_set)
+      {
+        //auto qq = *cut_entry;
+        const auto tt = cuts.truth_table( *cut_entry );
+        /* check if the function of the cut is NPN-equivalent to XOR3 */
+        if (std::find_if(xor3_tts.begin(), xor3_tts.end(), [&](const TT3 & xor3_tt){return tt._bits.front() == xor3_tt._bits;}) != xor3_tts.end())
+        {
+          std::array<uint32_t,3> item;
+          std::copy(cut_entry->begin(), cut_entry->end(), item.begin());
+          xor3_cuts.emplace( item, std::make_tuple(tt, idx) );
+
+          std::ostringstream oss;
+          oss << *cut_entry;
+          // fmt::print("[{}]\t XOR3 cut {} with tt 0b{}=0x{}\n", benchmark, oss.str(), kitty::to_binary( tt ), kitty::to_hex( tt ));
+        }
+        if (std::find_if(maj3_tts.begin(), maj3_tts.end(), [&](const TT3 & maj3_tt){return tt._bits.front() == maj3_tt._bits;}) != maj3_tts.end())
+        {
+          std::array<uint32_t,3> item;
+          std::copy(cut_entry->begin(), cut_entry->end(), item.begin());
+          maj3_cuts.emplace( item, std::make_tuple(tt, idx) );
+
+          std::ostringstream oss;
+          oss << *cut_entry;
+          // fmt::print("[{}]\t MAJ3 cut {} with tt 0b{}=0x{}\n", benchmark, oss.str(), kitty::to_binary( tt ), kitty::to_hex( tt ));
+        }
+        if ( std::find_if( or3_tts.begin(), or3_tts.end(), [&]( TT3 const& or3_tt ) { return tt._bits.front() == or3_tt._bits; } ) != or3_tts.end() )
+        {
+          std::array<uint32_t, 3> item;
+          std::copy( cut_entry->begin(), cut_entry->end(), item.begin() );
+          or3_cuts.emplace( item, std::make_tuple( tt, idx ) );
+        }
+        // std::ostringstream oss;
+        // oss << *cut_entry;
+        // fmt::print("[{}] \tCut {} with tt 0b{}=0x{}\n", benchmark, oss.str(), kitty::to_binary( tt ), kitty::to_hex( tt ));
+      }
+      // flat_hash_set<std::array<uint32_t, 3>> arraySet;
+      // auto tt = cuts.truth_table( node_cut_set );
+      // std::cout << node_cut_set << " -> " <<  << "\n";
+      // cuts.truth_table( node_cut_set );
+    } );
 
     /* TODO: adopt the assumption that, it is a good deal if an implementation can make use of more than 2 out of the 3 outputs of T1,  */
     /* then we would need four rounds of matching: (1) 3 leaves using all 3 outputs; (2) ...using XOR and MAJ; (3) ...using MAJ and OR; */
     /* (4) ...using XOR and OR. For each 3 leaves found, check validity by deciding which of the 8 T1s to choose                        */
-    array_map<3, T1_OUTPUTS> t1_candidates = find_t1_candidates(klut_decomposed, xor3_cuts, maj3_cuts, or3_cuts);
+    phmap::flat_hash_map<std::array<uint32_t,3>, T1_OUTPUTS, ArrayHash<3>> t1_candidates;
+    // std::vector<std::array<uint32_t, 3>> checked_leaves;
+
+    for ( auto const& xor3_cut : xor3_cuts )
+    {
+      auto target_leaves = xor3_cut.first;
+      if ( maj3_cuts.contains( target_leaves ) )
+      {
+        if ( or3_cuts.contains( target_leaves ) )
+        {
+          /* find a candidate potentially using all 3 outputs of an T1 cell */
+          /* check whether the combination of truth table is valid          */
+          bool maj_phase{ false }, or_phase{ false };
+          for ( auto const& t1_cell : standard_T1_cells )
+          {
+            maj_phase = or_phase = false;
+            if ( std::get<0>( xor3_cut.second )._bits[0] != t1_cell.sum_truth_table )
+            {
+              continue;
+            }
+
+            if ( std::get<0>( maj3_cuts[target_leaves] )._bits[0] == ~t1_cell.carry_truth_table )
+            {
+              maj_phase = true;
+            }
+            else if ( std::get<0>( maj3_cuts[target_leaves] )._bits[0] != t1_cell.carry_truth_table )
+            {
+              continue;
+            }
+
+            if ( std::get<0>( or3_cuts[target_leaves] )._bits[0] == ~t1_cell.cbar_truth_table )
+            {
+              or_phase = true;
+            }
+            else if ( std::get<0>( or3_cuts[target_leaves] )._bits[0] != t1_cell.cbar_truth_table )
+            {
+              continue;
+            }
+
+            /* managed to find a T1 cell-based implementation      */
+            /* figure out the nodes connecting to the output ports */
+            klut::node sum_to = klut_decomposed.index_to_node( std::get<1>( xor3_cut.second ) );
+            klut::node carry_to = klut_decomposed.get_constant( false );
+            klut::node inv_carry_to = klut_decomposed.get_constant( false );
+            klut::node cbar_to = klut_decomposed.get_constant( false );
+            klut::node inv_cbar_to = klut_decomposed.get_constant( false );
+            config_t1_ports_connection( klut_decomposed, maj_phase, std::get<1>( maj3_cuts[target_leaves] ), carry_to, inv_carry_to );
+            config_t1_ports_connection( klut_decomposed,  or_phase, std::get<1>(  or3_cuts[target_leaves] ),  cbar_to,  inv_cbar_to );
+
+            /* instantiate an T1 cell */
+            t1_candidates.emplace( target_leaves, T1_OUTPUTS( t1_cell.in_phase, inv_carry_to != klut_decomposed.get_constant( false ), inv_cbar_to != klut_decomposed.get_constant( false ), 
+                                                              sum_to != klut_decomposed.get_constant( false ), carry_to != klut_decomposed.get_constant( false ), cbar_to == klut_decomposed.get_constant( false ), 
+                                                              sum_to, carry_to, inv_carry_to, cbar_to, inv_cbar_to ) );
+            break;
+            /* TODO: think about if it is possible that a design can be implemented under */
+            /* different input phases of a T1 cell; if yes, it is better to enumerate all */
+            /* the potential input phases and choose the one requires the minimum cost    */
+          }
+        }
+      }
+    }
+
+    for ( auto const& xor3_cut : xor3_cuts )
+    {
+      auto target_leaves = xor3_cut.first;
+      if ( maj3_cuts.contains( target_leaves ) )
+      {
+        if ( !t1_candidates.contains( target_leaves ) )
+        {
+          /* find a candidate potentially using 2 outputs of an T1 cell  */
+          /* ( XOR3, MAJ3 ), check whether the combnation of tt is valid */
+          bool maj_phase{ false };
+          for ( auto const& t1_cell : standard_T1_cells )
+          {
+            if ( std::get<0>( xor3_cut.second )._bits[0] != t1_cell.sum_truth_table )
+            {
+              continue;
+            }
+
+            if ( std::get<0>( maj3_cuts[target_leaves] )._bits[0] == ~t1_cell.carry_truth_table )
+            {
+              maj_phase = true;
+            }
+            else if ( std::get<0>( maj3_cuts[target_leaves] )._bits[0] != t1_cell.carry_truth_table )
+            {
+              continue;
+            }
+
+            /* managed to find a T1 cell-based implementation      */
+            /* figure out the nodes connecting to the output ports */
+            klut::node sum_to = klut_decomposed.index_to_node( std::get<1>( xor3_cut.second ) );
+            klut::node carry_to = klut_decomposed.get_constant( false );
+            klut::node inv_carry_to = klut_decomposed.get_constant( false );
+            config_t1_ports_connection( klut_decomposed, maj_phase, std::get<1>( maj3_cuts[target_leaves] ), carry_to, inv_carry_to );
+
+            /* instantiate an T1 cell */
+            t1_candidates.emplace( target_leaves, T1_OUTPUTS( t1_cell.in_phase, inv_carry_to != klut_decomposed.get_constant( false ), false, 
+                                                              true, carry_to != klut_decomposed.get_constant( false ), false, 
+                                                              sum_to, carry_to, inv_carry_to, klut_decomposed.get_constant( false ), klut_decomposed.get_constant( false ) ) );
+            break;
+          }
+        }
+      }
+    }
+
+    for ( auto const& xor3_cut : xor3_cuts )
+    {
+      auto target_leaves = xor3_cut.first;
+      if ( or3_cuts.contains( target_leaves ) )
+      {
+        if ( !t1_candidates.contains( target_leaves ) )
+        {
+          /* find a candidate potentially using 2 outputs of an T1 cell  */
+          /* ( XOR3,  OR3 ), check whether the combnation of tt is valid */
+          bool or_phase{ false };
+          for ( auto const& t1_cell : standard_T1_cells )
+          {
+            if ( std::get<0>( xor3_cut.second )._bits[0] != t1_cell.sum_truth_table )
+            {
+              continue;
+            }
+
+            if ( std::get<0>( or3_cuts[target_leaves] )._bits[0] == ~t1_cell.cbar_truth_table )
+            {
+              or_phase = true;
+            }
+            else if ( std::get<0>( or3_cuts[target_leaves] )._bits[0] != t1_cell.cbar_truth_table )
+            {
+              continue;
+            }
+
+            /* managed to find a T1 cell-based implementation      */
+            /* figure out the nodes connecting to the output ports */
+            klut::node sum_to = klut_decomposed.index_to_node( std::get<1>( xor3_cut.second ) );
+            klut::node cbar_to = klut_decomposed.get_constant( false );
+            klut::node inv_cbar_to = klut_decomposed.get_constant( false );
+            config_t1_ports_connection( klut_decomposed, or_phase, std::get<1>( or3_cuts[target_leaves] ), cbar_to, inv_cbar_to );
+
+            /* instantiate an T1 cell */
+            t1_candidates.emplace( target_leaves, T1_OUTPUTS( t1_cell.in_phase, false, inv_cbar_to != klut_decomposed.get_constant( false ), 
+                                                              true, false, cbar_to != klut_decomposed.get_constant( false ), 
+                                                              sum_to, klut_decomposed.get_constant( false ), klut_decomposed.get_constant( false ), cbar_to, inv_cbar_to ) );
+            break;
+          }
+        }
+      }
+    }
+
+    for ( auto const& maj3_cut : maj3_cuts )
+    {
+      auto target_leaves = maj3_cut.first;
+      if ( or3_cuts.contains( maj3_cut.first ) )
+      {
+        if ( !t1_candidates.contains( target_leaves ) )
+        {
+          /* find a candidate potentially using 2 outputs of an T1 cell  */
+          /* ( MAJ3,  OR3 ), check whether the combnation of tt is valid */
+          bool maj_phase{ false }, or_phase{ false };
+          for ( auto const& t1_cell : standard_T1_cells )
+          {
+            if ( std::get<0>( maj3_cut.second )._bits[0] == ~t1_cell.carry_truth_table )
+            {
+              maj_phase = true;
+            }
+            else if ( std::get<0>( maj3_cut.second )._bits[0] != t1_cell.carry_truth_table )
+            {
+              continue;
+            }
+
+            if ( std::get<0>( or3_cuts[target_leaves] )._bits[0] == ~t1_cell.cbar_truth_table )
+            {
+              or_phase = true;
+            }
+            else if ( std::get<0>( or3_cuts[target_leaves] )._bits[0] != t1_cell.cbar_truth_table )
+            {
+              continue;
+            }
+
+            /* managed to find a T1 cell-based implementation      */
+            /* figure out the nodes connecting to the output ports */
+            klut::node carry_to = klut_decomposed.get_constant( false );
+            klut::node inv_carry_to = klut_decomposed.get_constant( false );
+            klut::node cbar_to = klut_decomposed.get_constant( false );
+            klut::node inv_cbar_to = klut_decomposed.get_constant( false );
+            config_t1_ports_connection( klut_decomposed, maj_phase, std::get<1>(          maj3_cut.second ), carry_to, inv_carry_to );
+            config_t1_ports_connection( klut_decomposed,  or_phase, std::get<1>(  or3_cuts[target_leaves] ),  cbar_to,  inv_cbar_to );
+
+            /* instantiate an T1 cell */
+            t1_candidates.emplace( target_leaves, T1_OUTPUTS( t1_cell.in_phase, inv_carry_to != klut_decomposed.get_constant( false ), inv_cbar_to != klut_decomposed.get_constant( false ), 
+                                                              false, carry_to != klut_decomposed.get_constant( false ), cbar_to != klut_decomposed.get_constant( false ), 
+                                                              klut_decomposed.get_constant( false ), carry_to, inv_carry_to, cbar_to, inv_cbar_to ) );
+            break;
+          }
+        }
+      }
+    }
 
     /* estimate the gain of implementing parts of the circuits using T1s instead */
     auto updated_area{ raw_area };
     // uint32_t num_t1_use_more_than_3{ 0u };
     // uint32_t num_t1_cells{ 0u };
-
-    /* Rewrote this using iterator output */
-    for ( auto it_t1_cands{ t1_candidates.begin() }; it_t1_cands != t1_candidates.end(); )
+    for ( auto it_t1_cands{ t1_candidates.begin() }; it_t1_cands != t1_candidates.end(); ++it_t1_cands )
     {
       if ( !t1_usage_sanity_check( klut_decomposed, *it_t1_cands, updated_area ) )
       {
-        // update iterator after erasing
         it_t1_cands = t1_candidates.erase( it_t1_cands );
       }
-      else
-      {
-        ++it_t1_cands;
-      }
+      // else
+      // {
+      //   ++num_t1_cells;
+      //   if ( customized_T1_classifier( ( *it_t1_cands ).second ) )
+      //   {
+      //     ++num_t1_use_more_than_3;
+      //   }
+      // }
+      // fmt::print( "Area before : {}, \tArea after : {}, \tRed. : {:>5.2f}%\n", raw_area, updated_area, ( ( static_cast<float>( raw_area ) - static_cast<float>( updated_area ) ) / static_cast<float>( raw_area ) * 100 ) );
     }
 
     /* bind the instantiated T1 cells to the 2-LUT network for csv generation    */
@@ -1557,6 +2438,7 @@ int main(int argc, char* argv[])  //
     // fmt::print( "\t# T1s : {}, # T1s with more than 3 ouput ports used : {}\n", num_t1_cells, num_t1_use_more_than_3 );
 
     fmt::print( "Area before : {}, \tArea after : {}, \tRed. : {:>5.2f}%\n", raw_area, updated_area, ( ( static_cast<float>( raw_area ) - static_cast<float>( updated_area ) ) / static_cast<float>( raw_area ) * 100 ) );
+    
     
 
     continue;
